Time: 2 ms (Beats 97.73%), Space: 48.4 MB (Beats 57.43%)

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.java
@@ -25,8 +25,8 @@ class Solution {
 
         while (!queue.isEmpty()) {
 
-            double levelSize = queue.size();
-            double sum = 0.0;
+            int levelSize = queue.size();
+            long sum = 0L;
             for (int i = 0; i < levelSize; i++) {
                 TreeNode node = queue.poll();
                 sum += node.val;
@@ -37,7 +37,7 @@ class Solution {
                     queue.offer(node.right);
                 }
             }
-            result.add(sum / levelSize);
+            result.add((double)sum / levelSize);
         }
         return result;
     }
